requirements.txt
# California Stewardship Fund — Intelligence Agents
# Python 3.10+ required

# HTTP client
requests>=2.31.0

# HTML parsing (used by leginfo scraper fallback)
beautifulsoup4>=4.12.0
lxml>=5.1.0

# YAML config files
PyYAML>=6.0.1

# Date parsing utilities
python-dateutil>=2.9.0

# .env file support for storing credentials
python-dotenv>=1.0.0

# Anthropic Claude SDK (used by housing_analyzer agent)
anthropic>=0.40.0

# RSS/Atom feed parsing (used by media_scanner agent)
feedparser>=6.0.11

# Google Gemini SDK (used by image_generator for Nano Banana Pro / gemini-3-pro-image-preview)
google-genai>=1.0.0

# PNG save from Gemini inline_data (used by image_generator)
Pillow>=10.0.0

# Testing
pytest>=8.0
pytest-mock>=3.12
freezegun>=1.4
